+++ api/src/controllers/router.js
@@ -1,5 +1,7 @@
 const Router = require('koa-router');
 const bodyParser = require('koa-bodyparser');
+const yamljs = require('yamljs');
+const { koaSwagger } = require('koa2-swagger-ui');
 const { authentify } = require('./auth');
 const { getVersion } = require('../services/version');
 const { getChallenge } = require('../services/auth');
@@ -40,10 +42,20 @@ const {
   addressOrAnySchema,
 } = require('../utils/validator');
 const { UNPUBLISH_TARGET_MAP } = require('../utils/order-utils');
+const { maxPageSize, minPageSize } = require('../config').api;
 
 const log = logger.extend('controllers:router');
 
 const router = new Router();
+
+// docs
+router.get(
+  '/docs',
+  koaSwagger({
+    routePrefix: false,
+    swaggerOptions: { spec: yamljs.load('./openapi.yaml') },
+  }),
+);
 
 // version
 router.get('/version', async (ctx) => {
@@ -98,20 +110,28 @@ router.get('/categories/:catid', bodyParser(), async (ctx) => {
 
 router.get('/categories', bodyParser(), async (ctx) => {
   log('GET /categories');
-  const { chainId, catid, minWorkClockTimeRef, maxWorkClockTimeRef, page } =
-    await object({
-      chainId: chainIdSchema().required(),
-      catid: positiveIntSchema(),
-      minWorkClockTimeRef: positiveIntSchema(),
-      maxWorkClockTimeRef: positiveIntSchema(),
-      page: positiveIntSchema(),
-    }).validate(ctx.query);
-  const { categories, count, nextPage } = await getCategories({
+  const {
     chainId,
-    catid,
     minWorkClockTimeRef,
     maxWorkClockTimeRef,
     page,
+    pageIndex,
+    pageSize,
+  } = await object({
+    chainId: chainIdSchema().required(),
+    minWorkClockTimeRef: positiveIntSchema(),
+    maxWorkClockTimeRef: positiveIntSchema(),
+    page: positiveIntSchema(),
+    pageIndex: positiveIntSchema(),
+    pageSize: positiveIntSchema().max(maxPageSize).min(minPageSize),
+  }).validate(ctx.query);
+  const { categories, count, nextPage } = await getCategories({
+    chainId,
+    minWorkClockTimeRef,
+    maxWorkClockTimeRef,
+    page,
+    pageIndex,
+    pageSize,
   });
   ctx.body = {
     ok: true,
@@ -160,6 +180,8 @@ router.get('/deals', bodyParser(), async (ctx) => {
     workerpoolorderHash,
     requestorderHash,
     page,
+    pageIndex,
+    pageSize,
   } = await object({
     chainId: chainIdSchema().required(),
     category: positiveIntSchema(),
@@ -176,6 +198,8 @@ router.get('/deals', bodyParser(), async (ctx) => {
     workerpoolorderHash: bytes32Schema(),
     requestorderHash: bytes32Schema(),
     page: positiveIntSchema(),
+    pageIndex: positiveIntSchema(),
+    pageSize: positiveIntSchema().max(maxPageSize).min(minPageSize),
   }).validate(ctx.query);
   const { deals, count, nextPage } = await getDeals({
     chainId,
@@ -193,6 +217,8 @@ router.get('/deals', bodyParser(), async (ctx) => {
     workerpoolorderHash,
     requestorderHash,
     page,
+    pageIndex,
+    pageSize,
   });
   ctx.body = {
     ok: true,
@@ -252,6 +278,8 @@ router.get('/apporders', bodyParser(), async (ctx) => {
     maxTag,
     minVolume,
     page,
+    pageIndex,
+    pageSize,
   } = await object({
     chainId: chainIdSchema().required(),
     app: string().when('appOwner', {
@@ -267,6 +295,8 @@ router.get('/apporders', bodyParser(), async (ctx) => {
     maxTag: bytes32Schema(),
     minVolume: positiveStrictIntSchema(),
     page: positiveIntSchema(),
+    pageIndex: positiveIntSchema(),
+    pageSize: positiveIntSchema().max(maxPageSize).min(minPageSize),
   }).validate(ctx.query);
   const { orders, count, nextPage } = await getApporders({
     chainId,
@@ -279,6 +309,8 @@ router.get('/apporders', bodyParser(), async (ctx) => {
     maxTag,
     minVolume,
     page,
+    pageIndex,
+    pageSize,
   });
   ctx.body = {
     ok: true,
@@ -327,6 +359,8 @@ router.get('/datasetorders', bodyParser(), async (ctx) => {
     maxTag,
     minVolume,
     page,
+    pageIndex,
+    pageSize,
   } = await object({
     chainId: chainIdSchema().required(),
     dataset: string().when('datasetOwner', {
@@ -344,6 +378,8 @@ router.get('/datasetorders', bodyParser(), async (ctx) => {
     maxTag: bytes32Schema(),
     minVolume: positiveStrictIntSchema(),
     page: positiveIntSchema(),
+    pageIndex: positiveIntSchema(),
+    pageSize: positiveIntSchema().max(maxPageSize).min(minPageSize),
   }).validate(ctx.query);
   const { orders, count, nextPage } = await getDatasetorders({
     chainId,
@@ -356,6 +392,8 @@ router.get('/datasetorders', bodyParser(), async (ctx) => {
     maxTag,
     minVolume,
     page,
+    pageIndex,
+    pageSize,
   });
   ctx.body = {
     ok: true,
@@ -406,6 +444,8 @@ router.get('/workerpoolorders', bodyParser(), async (ctx) => {
     minTrust,
     minVolume,
     page,
+    pageIndex,
+    pageSize,
   } = await object({
     chainId: chainIdSchema().required(),
     workerpool: addressOrAnySchema(),
@@ -419,6 +459,8 @@ router.get('/workerpoolorders', bodyParser(), async (ctx) => {
     minVolume: positiveStrictIntSchema(),
     minTrust: positiveIntSchema(),
     page: positiveIntSchema(),
+    pageIndex: positiveIntSchema(),
+    pageSize: positiveIntSchema().max(maxPageSize).min(minPageSize),
   }).validate(ctx.query);
   const { orders, count, nextPage } = await getWorkerpoolorders({
     chainId,
@@ -433,6 +475,8 @@ router.get('/workerpoolorders', bodyParser(), async (ctx) => {
     minTrust,
     minVolume,
     page,
+    pageIndex,
+    pageSize,
   });
   ctx.body = {
     ok: true,
@@ -483,6 +527,8 @@ router.get('/requestorders', bodyParser(), async (ctx) => {
     minVolume,
     workerpool,
     page,
+    pageIndex,
+    pageSize,
   } = await object({
     chainId: chainIdSchema().required(),
     app: addressOrAnySchema(),
@@ -496,6 +542,8 @@ router.get('/requestorders', bodyParser(), async (ctx) => {
     minVolume: positiveStrictIntSchema(),
     workerpool: addressOrAnySchema(),
     page: positiveIntSchema(),
+    pageIndex: positiveIntSchema(),
+    pageSize: positiveIntSchema().max(maxPageSize).min(minPageSize),
   }).validate(ctx.query);
   const { orders, count, nextPage } = await getRequestorders({
     chainId,
@@ -510,6 +558,8 @@ router.get('/requestorders', bodyParser(), async (ctx) => {
     minVolume,
     workerpool,
     page,
+    pageIndex,
+    pageSize,
   });
   ctx.body = {
     ok: true,